+++ data/src/main/java/com/daily/data/remote/datasource/auth/AuthDataSource.kt
@@ -1,5 +1,6 @@
 package com.daily.data.remote.datasource.auth
 
+import com.daily.data.remote.model.ChangePasswordRequestData
 import com.daily.data.remote.model.SignInRequestData
 import com.daily.data.remote.model.SignInResponseData
 import com.daily.data.remote.model.SignUpRequestData
@@ -12,4 +13,6 @@ interface AuthDataSource {
     suspend fun checkDuplicateEmail(email: String)
 
     suspend fun checkDuplicateName(name: String)
+
+    suspend fun changePassword(changePasswordRequestData: ChangePasswordRequestData)
 }
