+++ data/src/main/java/com/daily/data/repository/AccountRepositoryImpl.kt
@@ -1,8 +1,6 @@
 package com.daily.data.repository
 
 import com.daily.data.remote.datasource.account.AccountDataSource
-import com.daily.data.remote.model.asChangePasswordRequest
-import com.daily.domain.model.ChangePasswordRequest
 import com.daily.domain.repository.AccountRepository
 import javax.inject.Inject
 
@@ -11,9 +9,5 @@ class AccountRepositoryImpl @Inject constructor(
 ): AccountRepository {
     override suspend fun choiceTheme(theme: String) {
         accountDataSource.choiceTheme(theme)
-    }
-
-    override suspend fun changePassword(changePasswordRequest: ChangePasswordRequest) {
-        accountDataSource.changePassword(changePasswordRequest.asChangePasswordRequest())
     }
 }
