+++ domain/src/main/java/com/daily/domain/repository/AccountRepository.kt
@@ -1,9 +1,5 @@
 package com.daily.domain.repository
 
-import com.daily.domain.model.ChangePasswordRequest
-
 interface AccountRepository {
     suspend fun choiceTheme(theme: String)
-
-    suspend fun changePassword(changePasswordRequest: ChangePasswordRequest)
 }
