+++ domain/src/main/java/com/daily/domain/usecase/ChangePasswordUseCase.kt
@@ -1,11 +1,11 @@
 package com.daily.domain.usecase
 
 import com.daily.domain.model.ChangePasswordRequest
-import com.daily.domain.repository.AccountRepository
+import com.daily.domain.repository.AuthRepository
 import javax.inject.Inject
 
 class ChangePasswordUseCase @Inject constructor(
-    private val repository: AccountRepository
+    private val repository: AuthRepository
 ) {
     suspend operator fun invoke(changePasswordRequest: ChangePasswordRequest) = kotlin.runCatching { repository.changePassword(changePasswordRequest) }
 }
